Fix code review comments.
